requirements.txt
# FastAPI & Server
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.1

# Database
asyncpg==0.30.0
sqlalchemy[asyncio]==2.0.36
aiosqlite==0.22.1

# HTTP Client (for LLM + Auth calls)
httpx==0.28.1

# Security
python-jose[cryptography]==3.3.0

# Logging
structlog==24.4.0

# Testing
pytest==8.3.4
pytest-asyncio==0.25.0
pytest-cov==6.0.0

# Utilities
python-dotenv==1.0.1
